Support quoted literals. Simplify non-quoted literal rules.
Quoted literals allow the use of "" to specify literal boudaries. This
lets me parse literals in places where they would normally fail, and it
also supports terminating a literal and following it with a "<". (In
theory it could be followed by "<<", but it doesn't make sense for a
literal to take a body.)